pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keypulse"
version = "0.1.0"
description = "Local-first personal activity memory layer for macOS"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.7",
    "pydantic>=2.5",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Cocoa",
    "pyobjc-framework-AppKit",
    "pyobjc-framework-ApplicationServices",
    "pyobjc-framework-Quartz",
    "pyobjc-framework-Vision",
]

[project.scripts]
keypulse = "keypulse.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["keypulse*"]
